refactor: 출력함수 constants > main으로 이동
